+++ include/intersections.h
@@ -6,7 +6,7 @@
 /*   By: hsarhan <hsarhan@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/21 12:19:12 by mkhan             #+#    #+#             */
-/*   Updated: 2023/02/02 18:13:48 by hsarhan          ###   ########.fr       */
+/*   Updated: 2023/02/05 21:04:53 by hsarhan          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -34,7 +34,7 @@ struct s_ray
 typedef struct s_intersect		t_intersection;
 struct s_intersect
 {	
-	float		time;
+	double		time;
 	t_shape		*shape;
 	t_vector	point;
 	t_vector	normal;
@@ -61,17 +61,17 @@ bool			intersect(t_shape *shape, const t_ray *ray,
 
 t_intersection	*hit(t_intersections *xs);
 
-void			ray_position(t_vector *pos, const t_ray *ray, float time);
+void			ray_position(t_vector *pos, const t_ray *ray, double time);
 
 t_vector		normal_at(t_scene *scene, const t_shape *shape, const t_vector *itx_point);
 
-void			ray_from_cam(t_ray *ray, const t_camera *cam, float x, float y);
+void			ray_from_cam(t_ray *ray, const t_camera *cam, double x, double y);
 bool			intersect_cylinder(const t_ray *ray, t_shape *shape,
 				t_intersections *xs);
 bool		intersect_cone(const t_ray *ray, t_shape *shape, t_intersections *xs);
-void		spherical_map(float *u, float *v, t_vector *point);
-void		cylindrical_map(float *u, float *v, t_vector *point);
-void		cubical_map(float *u, float *v, t_vector *point);
+void		spherical_map(double *u, double *v, t_vector *point);
+void		cylindrical_map(double *u, double *v, t_vector *point);
+void		cubical_map(double *u, double *v, t_vector *point);
 bool		intersect_cube(t_shape *shape, t_ray *ray, t_intersections *xs);
 bool		test_box_axes(t_shape *b1, t_shape *b2, t_vector *resolution);
 t_vector	cube_normal(const t_shape *shape, const t_vector *itx_point);
