+++ include/render.h
@@ -6,7 +6,7 @@
 /*   By: hsarhan <hsarhan@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/18 11:29:17 by hsarhan           #+#    #+#             */
-/*   Updated: 2023/02/02 13:27:58 by hsarhan          ###   ########.fr       */
+/*   Updated: 2023/02/05 21:04:53 by hsarhan          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -36,7 +36,7 @@ void	prepare_computations(t_scene *scene, t_intersection *intersection,
 
 t_color	phong(t_intersection *itx, t_scene *scene, int light_idx);
 
-bool	is_shadowed(t_scene *scene, int light_idx, t_vector *itx_point, float *);
+bool	is_shadowed(t_scene *scene, int light_idx, t_vector *itx_point, double *);
 
 t_color	cast_reflection_ray(t_scene *scene, t_intersection *intersection,
 			int remaining, int light_idx);
@@ -46,8 +46,8 @@ int		get_color(t_worker *worker, int x, int y);
 void	set_color(t_worker *worker, int x, int y, int color);
 
 t_color	shade_point(t_intersections *arr, t_scene *scene, t_ray *ray);
-t_color	render_pixel(float x, float y, t_intersections *arr, t_worker *worker);
-t_color	super_sample_pixel(float x, float y, t_intersections *arr, t_worker *worker);
+t_color	render_pixel(double x, double y, t_intersections *arr, t_worker *worker);
+t_color	super_sample_pixel(double x, double y, t_intersections *arr, t_worker *worker);
 void	fill_in_horizontal(t_worker *worker, int threshold);
 void	fill_in_vertical(t_worker *worker, int threshold);
 void	update_loading_bar(t_worker *worker, int *line_counter);
